Disable all not-in-used plane once DRM Master is reset
Disable all not-in-used planed once DRM Master is reset.
To avoid that some surface will is committed by other APP
are remained.

Tests: Work well with Kmscube testing APP.
Tracked-On: None
